refactor: modularize api into lib
