+++ .gitignore
@@ -1,1 +1,2 @@
 .vercel
+node_modules/
