+++ lib/build.js
@@ -1,11 +1,7 @@
-// /api/build-card.js
-// Deterministic crawl → TradeCard JSON → (optional) push to WordPress using your schema:
-// 1) POST /wp-json/wp/v2/tradecard
-// 2) POST /wp-json/tradecard/v1/upload-image-from-url
-// 3) PATCH /wp-json/custom/v1/acf-sync/{id}
+// lib/build.js
+// Crawl a site and build a TradeCard structure from pages.
 
 const { scrapePage } = require('./scrape');
-const { inferTradecard } = require('./infer');
 
 const ALLOWED_PROTOCOLS = new Set(['http:', 'https:']);
 const isHttp = (u) => { try { return ALLOWED_PROTOCOLS.has(new URL(u).protocol); } catch { return false; } };
@@ -144,102 +140,4 @@ function buildTradecardFromPages(startUrl, pages) {
   };
 }
 
-// ----------------- WordPress (JWT) helpers per your OpenAPI schema -----------------
-
-const ENDPOINTS = {
-  createTradecard: (base) => `${base.replace(/\/$/, '')}/wp-json/wp/v2/tradecard`,
-  uploadImageFromUrl: (base) => `${base.replace(/\/$/, '')}/wp-json/tradecard/v1/upload-image-from-url`,
-  acfSync: (base, id) => `${base.replace(/\/$/, '')}/wp-json/custom/v1/acf-sync/${id}`
-};
-
-function bearerHeader() {
-  const token = process.env.WP_BEARER;
-  return token ? { Authorization: `Bearer ${token}` } : {};
-}
-
-async function wpCreateTradecard(base, title, status = 'draft') {
-  const res = await fetch(ENDPOINTS.createTradecard(base), {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json', ...bearerHeader() },
-    body: JSON.stringify({ title, status })
-  });
-  const json = await res.json().catch(() => ({}));
-  if (!res.ok) return { error: `createTradecard ${res.status}`, detail: json };
-  return json; // expect { id, ... }
-}
-
-async function wpUploadFromUrl(base, url) {
-  const res = await fetch(ENDPOINTS.uploadImageFromUrl(base), {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json', ...bearerHeader() },
-    body: JSON.stringify({ url })
-  });
-  const json = await res.json().catch(() => ({}));
-  if (!res.ok) return { error: `uploadImageFromUrl ${res.status}`, detail: json };
-  return json; // expect { id, url }
-}
-
-async function wpAcfSync(base, id, fields) {
-  const res = await fetch(ENDPOINTS.acfSync(base, id), {
-    method: 'PATCH',
-    headers: { 'Content-Type': 'application/json', ...bearerHeader() },
-    body: JSON.stringify(fields)
-  });
-  const json = await res.json().catch(() => ({}));
-  if (!res.ok) return { error: `acfSync ${res.status}`, detail: json };
-  return json;
-}
-
-// ---- Map TradeCard JSON → your flattened ACF fields (per your schema) ----
-// Only include keys that have values (your custom endpoint us
-
-exports.crawlSite = crawlSite;
-exports.buildTradecardFromPages = buildTradecardFromPages;
-
-exports.default = async function handler(req, res) {
-  const startUrl = req.query?.url;
-  if (!startUrl) return res.status(400).json({ error: 'Missing ?url=' });
-
-  const maxPages = Math.min(parseInt(req.query?.maxPages || '12', 10) || 12, 50);
-  const maxDepth = Math.min(parseInt(req.query?.maxDepth || '2', 10) || 2, 5);
-  const sameOriginOnly = (req.query?.sameOrigin ?? '1') !== '0';
-
-  try {
-    const pages = await crawlSite(startUrl, { maxPages, maxDepth, sameOriginOnly });
-    const result = buildTradecardFromPages(startUrl, pages);
-
-    if (req.query?.infer === '1' && process.env.OPENAI_API_KEY) {
-      try {
-        const inferred = await inferTradecard(result.tradecard);
-        const applied = [];
-        if (inferred?.business?.description) {
-          result.tradecard.business.description = inferred.business.description;
-          applied.push('business.description');
-        }
-        if (Array.isArray(inferred?.services?.list)) {
-          result.tradecard.services.list = inferred.services.list;
-          applied.push('services.list');
-        }
-        if (Array.isArray(inferred?.service_areas)) {
-          result.tradecard.service_areas = inferred.service_areas;
-          applied.push('service_areas');
-        }
-        if (inferred?.brand?.tone) {
-          result.tradecard.brand.tone = inferred.brand.tone;
-          applied.push('brand.tone');
-        }
-        if (Array.isArray(inferred?.testimonials)) {
-          result.tradecard.testimonials = inferred.testimonials;
-          applied.push('testimonials');
-        }
-        result.needs_inference = result.needs_inference.filter(k => !applied.includes(k));
-      } catch (err) {
-        console.warn('Inference failed:', err.message || err);
-      }
-    }
-
-    return res.status(200).json(result);
-  } catch (err) {
-    return res.status(500).json({ error: err.message || 'Failed to build card' });
-  }
-};
+module.exports = { crawlSite, buildTradecardFromPages };
