+++ lib/infer.js
@@ -1,4 +1,4 @@
-// api/infer.js
+// lib/infer.js
 // Calls OpenAI to infer additional TradeCard fields from scraped data.
 
 const fetch = require('node-fetch');
